src/roles/tradebot-poc/requirements.txt
click
numpy
pandas
rich
yfinance
plotly
matplotlib
plotly
asyncio
trendln
python-binance
art
TA-lib
ruamel.yaml